+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 -   Added `chevronColor` property onto `<InfoListItem>`.
--   Added `chevronIcon` class for `chevron` element in `<InfoListItem>`.
+-   Added `chevron` class onto `<InfoListItem>`.
 -   Added 1rem default padding to the root styles of `<EmptyState>`. ([#320](https://github.com/brightlayer-ui/react-component-library/issues/320))
 
 ## v5.3.3 (December 6, 2021)